backports: add backports-savedefconfig rule and per-platform defconfigs.
The earlier package always used the same config on all platforms, which
didn't make sense once we were using more than one platform.  We also add a
'savedefconfig' option (which works like the one in buildroot and newer
kernels) to minimize the content of the defconfig.

The original config is no longer relevant.  We'll add back platform-specific
ones in subsequent patches.

Change-Id: Ic9205c188c42f17bb7c52ddd253122bc78dbe608