is-storage-box: return true if /rw/force-storage-box exists.
This allows you to tell a TV box to pretend to be a storage box, and thus
run the "full" version of SageTV instead of the client-only version.  It
seems to work, though you also have to get some storage from somewhere
(perhaps a USB drive mounted from startupmods).

Relatedly, changed S95sageserver to check is-storage-box before is-tv-box.
The former is a superset of functionality of the latter.

Change-Id: I577192b1e8a43dc1cff57927e6f83a632da35fe3