backports: mwifiex: stop individual BSSes at shutdown time.
This shouldn't be necessary; CMD_FUNC_SHUTDOWN is supposed to do this
automatically.  But adding this greatly increases the probability that
CMD_FUNC_INIT will succeed on the next load of the driver.

This symptom specifically occurs if there is a nonzero number of stations
connected at driver unload time.  If there aren't any stations connected,
the problem does not occur.

Also, run CMD_FUNC_SHUTDOWN for CMD_FUNC_INIT fails.  This might help
prevent the firmware from continuing to advertise its SSID when it's broken.

Change-Id: I5998b871b465f57102eb9e9f30a8729d7a028503