BlueZ: Adjust init script
Experimentation shows that Marvell 8897 BT adapter is sometimes still in INIT
state when we are about to reset it and change the adapter MAC address.
This leads to unstable adapter behavior. Add a check for INIT state and delay
extra 1 second if needed. Also, move the adapter MAC address change and resets
to background subshell, so that they don't affect the overall bootup time.

Change-Id: I89cf286b978e7af9eb21bd6a3d4b32df952f9936