'S80dnsmasq stop' didn't actually kill the dnsmasq babysitter.
The result was that when reconfiguring dnsmasq, you'd end up with multiple
copies of the babysitter running, and thus lots of log messages about how
dnsmasq kept dying (because the undead babysitter was trying to start new
copies while another dnsmasq was already running).

While we're here, remove unnecessary use of eval.  That's not what it's for.

Also redirect dnsmasq stderr to logos.

Change-Id: Ie65c01772aad6aafd567db7221c24c5f2683d24e