ath10k: add an mb() in the transmit function to fix firmware crashes.
To be honest, I have no idea why this helps.  I was desperate so I started
with udelay(1000), which fixed the crashes, then worked my way down to
ndelay(1) and it still fixed the crashes.  Even mb() fixes the crashes.
Removing it brings the crashes back instantly.  So this matters.

Change-Id: If57021435934c485c9a50f9b794648c78cf139a3