Make .stamp_downloaded useful for repo-bound packages
For packages with a repo-bound download URL, the only way to trigger a
"download" of the latest code (i.e., git-new-workdir a copy from
vendor/google/xyz into out/build/google_xyz directory) is by removing
the build directory for this project ("make xyz-dirclean"). Removing
just the .stamp_downloaded doesn't do anything because we skip the
download if the build directory exists (one reason for this is that
git-new-workdir doesn't like to work with existing directories).

The fix is to handle the case of an existing build directory and
simply do a "git reset --hard" to pick up the latest code from the
local git repository.

With this change one can actually rebuild a project by simply removing
the ".stamp_downloaded" and then doing a "make google_xyz" which
should speed up things.

Change-Id: I1a561ac07a6127be7aa6c0e4ccaa213d3502b44b