rg_qa_settings -> rg-qa-settings
The main reason for this rename is that there are devices in the field that
are calling this script from /rw/startupmods.  That mechanism is obsolete
and causes conflict with ACS-based configuration.  So rename this script to
fix up those boxes, without preventing people from actually still running
this script from the command line if they want.

Change-Id: I266b8479e0a0e3306d53f4b3fef0d435dbdb58af