Drop openntpd, use tlsdate.
tlsdate is what ChromeOS uses, and it's hard to argue with the logic:
- actual secure verification of timestamps using TLS/SSL
- no strange side effects from miscalculated adjtime()
- generally simpler code

The down side is tlsdate doesn't know how to use adjtime() to adjust for
clock rate skew.  So you might get occasional clock jumps instead of actual
rate adjustments.  The up side is openntpd was doing a pretty bad job of
that anyway on our platform, so actually this makes our clock
generally *more* accurate (both value and rate) than it was.

While we're here, fix a stray reference to zsh and add 'randint' as a clean
way of generating random numbers.

Also, don't bother to restart the time service anymore when dhcp
configuration changes; tlsdate-routeup is a handy tool that prints a line to
stdout whenever a new network route comes up, so we can auto-rerun
tlsdate automatically when that happens.

Change-Id: I13f38507ec2df461a9ab1764c4eba342f16354c1