Removed FFMPEG from 2.0 RG build
This fix has the effect of removing /app/sage/ffmpeg and
/app/client/libPullreader.so from the RG 2.0 build. Other changes
were checked in that removes the need for FFMPEG on the server.
We are keeping it on the 1.0 hardware since we are already paying
the licensing on those units that are manufactured. We also are
keeping it on both TV Boxes since now FFMPEG is invoked over the
network for doing format detection and we are paying for the
codec licensing on both versions of the TV Box.

Change-Id: If16246be6dbd474e925956892335a65a3ec33550